ADD: Subway Migration (part 1)
Based on the existing tables, create the migration script.
This script creates the static tables for subway orders and pre-filles
those.
This script creates the meal and order table.

FIX: update the fieldnames on the order table: add '_id' to every
referencing column.
The modification on the model reflects that change.

WIP: The views still need to be adjusted. Those views show only the
numerical representation of the choices. Those need to be preloaded
relations (string representation) or preloaded lookup tables (drop down
select presentation).
That lies out of the scope of this commit, though.